Make chunk coordinate type 32 bits
This ensures there isn't a problem converting between world coords
and chunk coords.

Since the world center if half of max of u32, this is a
very real problem to handle with only 16 bits.

+++ src/world.rs
@@ -1,7 +1,7 @@
 use fnv::FnvHashMap;
 use rand::{thread_rng, Rng};
 
-type ChunkPosScalar = u16;
+type ChunkPosScalar = u32;
 
 #[derive(Hash, PartialEq, Eq, Debug, Clone, Copy)]
 pub struct ChunkPos {
